backend/requirements.txt
Django==4.0.0
djangorestframework==3.13.0
django-cors-headers==3.11.0
django-filter==21.1
django-extensions==3.1.1
psycopg2-binary==2.9.1
redis==4.1.0
celery==5.2.0
gunicorn==20.1.0
python-dotenv==0.19.0
pydantic==1.9.0
requests==2.27.0
pytest==6.2.5
pytest-django==4.5.2
pytest-cov==3.0.0
factory-boy==3.2.1
faker==10.0.0
pyclamav==0.4.0
boto3==1.20.0
